wip: sync dev server to android
