+++ src/gears/Cargo.toml
@@ -16,10 +16,6 @@ objc2 = { version = "0.5.2", optional = true }
 anyhow = "1.0.86"
 serde = {version = "1.0.204", features = ["derive"]}
 serde_json = "1.0.120"
-reqwest = { version = "0.11", features = ["json"] }
-tokio = { version = "1", features = ["full"] }
-tokio-tungstenite = "0.15"
-futures-util = "0.3"
 regex = "1.7.0"
 
 [dependencies.uuid]
