+++ src/gears/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod android;
 
+// pub mod dev_server;
 #[cfg(feature = "ios")]
 pub mod ios;
 pub mod lua;
